Remove useStripeSubscription hook
