+++ packages/front-end/services/UserContext.tsx
@@ -34,6 +34,7 @@ import { GROWTHBOOK_SECURE_ATTRIBUTE_SALT } from "shared/constants";
 import { Permissions, userHasPermission } from "shared/permissions";
 import { getValidDate } from "shared/dates";
 import sha256 from "crypto-js/sha256";
+import { useFeature } from "@growthbook/growthbook-react";
 import {
   getGrowthBookBuild,
   getSuperadminDefaultRole,
@@ -118,6 +119,8 @@ export interface UserContextValue {
     experiments: string[];
     features: string[];
   };
+  canSubscribe: boolean;
+  freeSeats: number;
 }
 
 interface UserResponse {
@@ -163,6 +166,8 @@ export const UserContext = createContext<UserContextValue>({
     experiments: [],
     features: [],
   },
+  canSubscribe: false,
+  freeSeats: 3,
 });
 
 export function useUser() {
@@ -182,6 +187,8 @@ export function getCurrentUser() {
 
 export function UserContextProvider({ children }: { children: ReactNode }) {
   const { isAuthenticated, orgId, setOrganizations } = useAuth();
+
+  const selfServePricingEnabled = useFeature("self-serve-billing").on;
 
   const { data, mutate: mutateUser, error } = useApi<UserResponse>(`/user`, {
     shouldRun: () => isAuthenticated,
@@ -423,6 +430,37 @@ export function UserContextProvider({ children }: { children: ReactNode }) {
     if (data) setReady(true);
   }, [data]);
 
+  const organization: Partial<OrganizationInterface> | undefined =
+    currentOrg?.organization;
+  const subscription = currentOrg?.subscription || null;
+  const license = currentOrg?.license;
+
+  const canSubscribe = useMemo(() => {
+    const disableSelfServeBilling =
+      organization?.disableSelfServeBilling || false;
+
+    if (disableSelfServeBilling) return false;
+
+    if (organization?.enterprise) return false; //TODO: Remove this once we have moved the license off the organization
+
+    if (license?.plan === "enterprise") return false;
+
+    // if already on pro, they must have a stripeSubscription - some self-hosted pro have an annual contract not directly through stripe.
+    if (
+      license &&
+      ["pro", "pro_sso"].includes(license.plan || "") &&
+      !subscription?.externalId
+    )
+      return false;
+
+    if (!selfServePricingEnabled) return false;
+
+    if (["active", "trialing", "past_due"].includes(subscription?.status || ""))
+      return false;
+
+    return true;
+  }, [organization, license, subscription, selfServePricingEnabled]);
+
   return (
     <UserContext.Provider
       value={{
@@ -440,8 +478,8 @@ export function UserContextProvider({ children }: { children: ReactNode }) {
         permissions,
         permissionsUtil,
         settings: currentOrg?.organization?.settings || {},
-        license: currentOrg?.license,
-        subscription: currentOrg?.subscription || null,
+        license,
+        subscription,
         enterpriseSSO: currentOrg?.enterpriseSSO || undefined,
         accountPlan: currentOrg?.accountPlan,
         effectiveAccountPlan: currentOrg?.effectiveAccountPlan,
@@ -449,12 +487,14 @@ export function UserContextProvider({ children }: { children: ReactNode }) {
         licenseError: currentOrg?.licenseError || "",
         commercialFeatures: currentOrg?.commercialFeatures || [],
         apiKeys: currentOrg?.apiKeys || [],
-        organization: currentOrg?.organization || {},
+        organization: organization || {},
         seatsInUse: currentOrg?.seatsInUse || 0,
         teams,
         error: error?.message || orgLoadingError?.message,
         hasCommercialFeature: (feature) => commercialFeatures.has(feature),
         watching: watching,
+        canSubscribe,
+        freeSeats: organization?.freeSeats || 3,
       }}
     >
       {children}
